fix: improve accuracy of gamma and ln_gamma

+++ src/generic/exp10.rs
@@ -66,7 +66,7 @@ fn exp10_split<F: Exp10>(x: F) -> (i32, F, F) {
     let r_hi = t_hi * F::LN_10_HI;
     let r_lo = t_hi * F::LN_10_LO + t_lo * F::LN_10;
 
-    (k, r_hi, r_lo)
+    (k as i32, r_hi, r_lo)
 }
 
 #[cfg(test)]
